Move tests into TableRegistry suite as they belong there.
